+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/snapshot/SnapshotCheckProcess.java
@@ -724,8 +724,6 @@ public class SnapshotCheckProcess {
 
     /** */
     private void registerMetrics(SnapshotCheckContext ctx) {
-        log.error("TEST | registerMetrics");
-
         MetricRegistryImpl mreg = kctx.metric().registry(MetricUtils.metricName(SNAPSHOT_CHECK_METRIC, ctx.req.snapshotName()));
 
         assert !mreg.iterator().hasNext();
